@@ -11,42 +11,24 @@ namespace Structural\Adapter;
 
 class Registry implements RegistryInterface
 {
-    /**
-     * @var string
-     */
-    private $first;
-    /**
-     * @var string
-     */
-    private $second;
+    private string $first;
+    private string $second;
 
-    /**
-     * @param string $value
-     */
     public function setFirst(string $value)
     {
         $this->first = $value;
     }
 
-    /**
-     * @param string $value
-     */
     public function setSecond(string $value)
     {
         $this->second = $value;
     }
 
-    /**
-     * @return string
-     */
     public function getFirst(): string
     {
         return $this->first;
     }
 
-    /**
-     * @return string
-     */
     public function getSecond(): string
     {
         return $this->second;
